Bugfix: make fdso syscalls 64bit clean
This fixes a bug found by liblitmus's regression test suite.
Before:
> ** LITMUS^RT test suite.
> ** Running tests for LINUX.
> ** Testing: don't open FMLP semaphores if FMLP is not supported...
> !! TEST FAILURE open_fmlp_sem(fd, 0) -> -16, Success (expected: EBUSY)
>    at tests/fdso.c:21 (test_fmlp_not_active)
> ** Testing: reject invalid object descriptors... ok.
> ** Testing: reject invalid object types...
> !! TEST FAILURE od_open(0, -1, 0) -> -22, Bad file descriptor (expected: EINVAL)
>    at tests/fdso.c:51 (test_invalid_obj_type)
> ** Testing: reject invalid rt_task pointers... ok.
> ** Result: 2 ok, 2 failed.

After:
> ** LITMUS^RT test suite.
> ** Running tests for LINUX.
> ** Testing: don't open FMLP semaphores if FMLP is not supported... ok.
> ** Testing: reject invalid object descriptors... ok.
> ** Testing: reject invalid object types... ok.
> ** Testing: reject invalid rt_task pointers... ok.
> ** Result: 4 ok, 0 failed.

+++ litmus/fdso.c
@@ -223,7 +223,7 @@ struct od_table_entry* __od_lookup(int od)
 }
 
 
-asmlinkage int sys_od_open(int fd, int type, int obj_id, void* __user config)
+asmlinkage long sys_od_open(int fd, int type, int obj_id, void* __user config)
 {
 	int ret = 0;
 	struct file*  file;
@@ -259,7 +259,7 @@ out:
 }
 
 
-asmlinkage int sys_od_close(int od)
+asmlinkage long sys_od_close(int od)
 {
 	int ret = -EINVAL;
 	struct task_struct *t = current;
